изменено:      Android/BluZ_v1/app/src/main/java/ru/starline/bluz/MainActivity.kt 	изменено:      Android/BluZ_v1/app/src/main/java/ru/starline/bluz/NumberFragment.kt 	изменено:      Android/BluZ_v1/app/src/main/java/ru/starline/bluz/globalObj.kt 	изменено:      Android/BluZ_v1/app/src/main/res/layout/setup_layout.xml 	изменено:      Android/BluZ_v1/app/src/main/res/layout/spectr_layout.xml 	изменено:      Android/BluZ_v1/app/src/main/res/values/strings.xml 	изменено:      stm32cubeide/BluZ_v1/Core/Inc/main.h 	изменено:      stm32cubeide/BluZ_v1/Release/BluZ_v1.hex 	изменено:      stm32cubeide/BluZ_v1/Release/BluZ_v1.list 	изменено:      stm32cubeide/BluZ_v1/Release/BluZ_v1.map 	изменено:      stm32cubeide/BluZ_v1/Release/STM32_WPAN/App/bluz.o 	изменено:      stm32cubeide/BluZ_v1/Release/STM32_WPAN/App/bluz_app.cyclo 	изменено:      stm32cubeide/BluZ_v1/Release/STM32_WPAN/App/bluz_app.o 	изменено:      stm32cubeide/BluZ_v1/Release/STM32_WPAN/App/bluz_app.su 	изменено:      stm32cubeide/BluZ_v1/STM32_WPAN/App/bluz.c 	изменено:      stm32cubeide/BluZ_v1/STM32_WPAN/App/bluz_app.c

+++ Android/BluZ_v1/app/src/main/java/ru/starline/bluz/MainActivity.kt
@@ -196,8 +196,8 @@ public class MainActivity : FragmentActivity() {
         GO.propCoefC = GO.PP.getPropFloat(propCoefC)
         GO.propCPS2UR = GO.PP.getPropFloat(propCPS2UR)
         GO.propIndicator = GO.PP.getPropByte(propIndicator)
-        GO.propHVoltage = GO.PP.getPropInt(propHV).toUInt()
-        GO.propComparator = GO.PP.getPropInt(propComparator).toUInt()
+        GO.propHVoltage = GO.PP.getPropInt(propHV).toUShort()
+        GO.propComparator = GO.PP.getPropInt(propComparator).toUShort()
 
         Log.d("BluZ-BT", "mac addr: " + GO.LEMAC)
     }
